Fix bug: inherited attributes incorrectly prioritized in RadialGradientElements
Per https://www.w3.org/TR/SVG11/pservers.html#RadialGradientElementFXAttribute, the fx coordinate is one of:
* defined on this element
* defaulted to the value of cx
* inherited from the referenced element

The prioritization of inheritance vs "coinciding" is somewhat ambiguous in this portion of the spec. The xlink attribute section (https://www.w3.org/TR/SVG11/pservers.html#RadialGradientElementHrefAttribute) clarifies somewhat: any attributes that are defined on the referenced element, but not on this element, are inherited - implying that the cx fallback should only be used in the absence of a referenced element.